TsumBeta: Add wait times on game start
Often the script on game start "thinks" that it is still on the StartScreen page and tries to check bonus items. This does not work as it is the wrong page, and leads to around 10 seconds at game start where no Tsums are connected.

These sleeps should make the StartPage to "any page" behaviour slower, but more reliable.