TsumBeta: Improve HTML Excel export of recorded data
The export file which contained two html tables did not contain any styles and the per-day results of the records were not structured well.

This reworked export creates separate columns for separate dates, inserts the per-record data into the correct columns and also enhances the visual representation of the table by adding table borders and different colors for different columns.